Add files via upload
"ReneWind” A renewable energy sources play an increasingly important role in the global energy mix, as the effort to reduce the environmental impact of energy production increases. They have shared a ciphered version of the data, as the data collected through sensors is confidential (the type of data collected varies with companies).
The objective is to build various classification models, tune them, and find the best one to help identify failures so that the generators can be repaired before failing/breaking to reduce the overall maintenance cost.